+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.5.0
+ Sept. 02, 2019
+- import 'package:flutter_localizations/flutter_localizations.dart';  
+- static final List<Locale> supportedLocales =
+- Add onGenerateTitle: 
+
 ## 0.4.1
  Sept. 01, 2019
 - factory I18nDelegate